% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eng_rb.R
\name{eng_rb}
\alias{eng_rb}
\title{A \code{knitr} engine for RevBayes}
\usage{
eng_rb(options)
}
\arguments{
\item{options}{Variables that can be specified within \code{knitr} chunk
headers to control how RevBayes code is executed. 
Those relevant to use \code{eng_rb} can be broken into three groups:
}
\emph{Generic Options}
\describe{
\item{options$code}{A string; the code for that chunk.}
\item{options$error}{A logical; should knitr fail on an error.}
\item{options$eval}{A logical; should the code be evaluated.}
\emph{Generic Options to specify when using RevBayes}
\describe{
\item{options$engine}{A string; Should be 'rb' to use RevBayes}
\item{options$engine.path}{A string; The path to the rb executable if not in the PATH}
\item{options$engine.opts}{A string; Command-line modifications for when rb is executed}
}
\emph{options for specifically for revbayes}
\describe{
\item{options$rbHistoryDirPath}{A string giving the path to the
directory where history files will be stored as RevBayes code is crunched.
The default path is ".eng_rb.knitr.cache" in the current working directory.
}
\item{options$refreshHistoryRB}{A logical. If not defined, the default
is \code{TRUE}. If \code{options$refreshHistoryRB = TRUE} and the hidden
chunk count function suggests this is the first RevBayes chunk
of the \code{knitr} document being knitted, then existing history files
found in \code{options$rbHistoryDirPath} from previous RevBayes analyses
using \code{knitr} will be removed. 
}
\item{options$rbDiagnosticMode}{A logical; default is \code{FALSE}. 
This option controlls whether to run diagnostic mode to
identify issues with the RevBayes engine.
}
}}
}
\value{
Output from applying RevBayes to the input code.
}
\description{
This function provides users with a way to execute code written
from the Bayesian phylogenetic program RevBayes as knitted chunks 
in a document, such as a chunk in a n ]code{rmarkdown} chunk.
}
\details{
There are a number of limitations of RevBayes that make this execution difficult:
there is presently no way to save environments and reload them across RevBayes instances,
and no apparent way to keep a RevBayes session open while getting the \code{stdout}, and
telling R that the process has terminated.
}
\examples{
# no examples yet
}
\seealso{
\code{\link{knitr::engine_output}}
}
\author{
April Wright, David W. Bapst
}
